refactor(ui5-file-uploader): migrate to TS
