+++ packages/main/src/features/InputElementsFormSupport.ts
@@ -1,14 +1,27 @@
 import { registerFeature } from "@ui5/webcomponents-base/dist/FeaturesRegistry.js";
+import type UI5Element from "@ui5/webcomponents-base/dist/UI5Element";
+
+interface IFormElement extends UI5Element {
+	value: string,
+	name: string,
+	disabled: boolean,
+	required?: boolean,
+	multiple?: boolean,
+}
+
+type NativeInputUpdateCallback = (element: IFormElement, nativeInput: HTMLInputElement) => void;
+type NativeInputChangeCallback = (e: Event) => void;
 
 class FormSupport {
 	/**
-	 *
-	 * @param element - the WebComponent that needs form support
-	 * @param nativeInputUpdateCallback - determines how the native input's disabled and value properties are calculated
+	 * Syncs the native input element, rendered into the component's light DOM,
+	 * with the component's state.
+	 * @param { IFormElement} element - the component with form support
+	 * @param { NativeInputUpdateCallback } nativeInputUpdateCallback - callback to calculate the native input's "disabled" and "value" properties
 	 */
-	static syncNativeHiddenInput(element, nativeInputUpdateCallback) {
+	static syncNativeHiddenInput(element: IFormElement, nativeInputUpdateCallback?: NativeInputUpdateCallback) {
 		const needsNativeInput = !!element.name || element.required;
-		let nativeInput = element.querySelector("input[data-ui5-form-support]");
+		let nativeInput = element.querySelector("input[data-ui5-form-support]") as HTMLInputElement;
 		if (needsNativeInput && !nativeInput) {
 			nativeInput = document.createElement("input");
 
@@ -21,11 +34,11 @@ class FormSupport {
 			nativeInput.style.width = "1px";
 			nativeInput.style.bottom = "0";
 			nativeInput.setAttribute("tabindex", "-1");
-			nativeInput.required = element.required;
+			nativeInput.required = element.required || false;
 			nativeInput.setAttribute("data-ui5-form-support", "");
 			nativeInput.setAttribute("aria-hidden", "true");
 
-			nativeInput.addEventListener("focusin", event => element.getFocusDomRef().focus());
+			nativeInput.addEventListener("focusin", () => element.getFocusDomRef()?.focus());
 
 			nativeInput.slot = "formSupport"; // Needed for IE - otherwise input elements are not part of the real DOM tree and are not detected by forms
 			element.appendChild(nativeInput);
@@ -40,13 +53,20 @@ class FormSupport {
 		}
 	}
 
-	static syncNativeFileInput(element, nativeInputUpdateCallback, nativeInputChangeCallback) {
+	/**
+	 * Syncs the native file input element, rendered into the <code>ui5-file-uploader</code> component's light DOM,
+	 * with the <code>ui5-file-uploader</code> component's state.
+	 * @param { IFormFileElement} element - the component with form support
+	 * @param { NativeInputUpdateCallback } nativeInputUpdateCallback - callback to calculate the native input's "disabled" and "value" properties
+	 * @param { NativeInputChangeCallback } nativeInputChangeCallback - callback, added to native input's "change" event
+	 */
+	static syncNativeFileInput(element: IFormElement, nativeInputUpdateCallback: NativeInputUpdateCallback, nativeInputChangeCallback: NativeInputChangeCallback) {
 		const needsNativeInput = !!element.name;
-		let nativeInput = element.querySelector(`input[type=${element._type || "hidden"}][data-ui5-form-support]`);
+		let nativeInput = element.querySelector(`input[type="file"][data-ui5-form-support]`) as HTMLInputElement;
 
 		if (needsNativeInput && !nativeInput) {
 			nativeInput = document.createElement("input");
-			nativeInput.type = element._type;
+			nativeInput.type = "file";
 			nativeInput.setAttribute("data-ui5-form-support", "");
 			nativeInput.slot = "formSupport"; // Needed to visualize the input in the light dom
 			nativeInput.style.position = "absolute";
@@ -75,13 +95,13 @@ class FormSupport {
 		}
 	}
 
-	static triggerFormSubmit(element) {
+	static triggerFormSubmit(element: UI5Element) {
 		let currentElement = element.parentElement;
 		while (currentElement && currentElement.tagName.toLowerCase() !== "form") {
 			currentElement = currentElement.parentElement;
 		}
 
-		if (currentElement) {
+		if (currentElement instanceof HTMLFormElement) {
 			if (!currentElement.checkValidity()) {
 				currentElement.reportValidity();
 				return;
@@ -103,7 +123,7 @@ class FormSupport {
 	}
 }
 
-const copyDefaultProperties = (element, nativeInput) => {
+const copyDefaultProperties = (element: IFormElement, nativeInput: HTMLInputElement) => {
 	nativeInput.disabled = element.disabled;
 	nativeInput.value = element.value;
 };
@@ -112,3 +132,9 @@ const copyDefaultProperties = (element, nativeInput) => {
 registerFeature("FormSupport", FormSupport);
 
 export default FormSupport;
+
+export {
+	IFormElement,
+	NativeInputChangeCallback,
+	NativeInputUpdateCallback,
+};
